Extract utility functions from ABI encoder.
